requirements.txt
#python = 3.10.16
#pip = 24.3.1 

annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
distro==1.9.0
elevenlabs==1.50.4
exceptiongroup==1.2.2
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SpeechRecognition==3.14.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
